Create Projects-by-last-edit
This is a DataviewJS query that returns the ten projects I most recently edited.

It's normally very easy to return a list of recently edited pages, but I wanted to do something harder—I wanted to extend the idea of "edited" to include edits to subpages, which are loosely defined through linkages and tags. This code searches through all pages to find the subprojects of a project page, and associates the project page with the most recent modification time of any subpage.